feature/be/#18 feat: 매니저의 매칭 확인 기능 추가
매니저가 자신에게 온 매칭을 확인 할 수 있는 api추가

+++ api/src/main/java/kernel/maidlab/api/reservation/service/ReservationServiceImpl.java
@@ -102,7 +102,6 @@ public class ReservationServiceImpl implements ReservationService {
 			// TODO : 수요자에게 알림 보내기 (예약 성공)
 		} else {
 			matchingService.changeStatus(reservationId, Status.REJECTED);
-
 		}
 	}
 
@@ -163,7 +162,9 @@ public class ReservationServiceImpl implements ReservationService {
 		}
 		reservation.cancel(LocalDateTime.now());
 		reservationRepository.save(reservation);
-		matchingRepository.deleteById(matchingRepository.findByReservationId(reservationId).getId());
+		if(matchingRepository.existsById(matchingRepository.findByReservationId(reservationId).getId())) {
+			matchingRepository.deleteById(matchingRepository.findByReservationId(reservationId).getId());
+		}
 	}
 
 	@Override
